feat(api): Made mustache & sprintf are now peers
BREAKING CHANGE: To reduce the weight of the package mustache and sprintf are now peer dependencies.
