+++ core.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.export = require('./lib/createRoot.js')
